# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BIMLog API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/submittals:
    get:
      operationId: listSubmittals
      tags: [submittals]
      summary: List submittals for a project
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submittal list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submittal"
    post:
      operationId: createSubmittal
      tags: [submittals]
      summary: Create a submittal
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmittalRequest"
      responses:
        "201":
          description: Submittal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submittal"
  /projects/{projectId}/submittals/{submittalId}:
    patch:
      operationId: updateSubmittal
      tags: [submittals]
      summary: Update a submittal
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: submittalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubmittalRequest"
      responses:
        "200":
          description: Submittal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submittal"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Submittal:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        number:
          type: string
        title:
          type: string
        description:
          type: string
        status:
          type: string
        specSection:
          type: string
        submittalType:
          type: string
        submittedById:
          type: integer
        submittedByName:
          type: string
        assignedToId:
          type: integer
        assignedToName:
          type: string
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - number
        - title
        - status
        - submittedById
        - submittedByName
        - createdAt
        - updatedAt
    CreateSubmittalRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        specSection:
          type: string
        submittalType:
          type: string
          enum: [shop_drawing, product_data, sample, mock_up, design_data, test_report, certificate, manufacturers_instructions, other]
        assignedToId:
          type: integer
        dueDate:
          type: string
          format: date-time
      required:
        - title
        - submittalType
    UpdateSubmittalRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, submitted, approved, approved_as_noted, rejected, resubmit, closed]
        specSection:
          type: string
        assignedToId:
          type: integer
        dueDate:
          type: string
          format: date-time
...