# lib/api-zod/src/generated/types/submittal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BIMLog API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Submittal {
  id: number;
  projectId: number;
  number: string;
  title: string;
  description?: string;
  status: string;
  specSection?: string;
  submittalType?: string;
  submittedById: number;
  submittedByName: string;
  assignedToId?: number;
  assignedToName?: string;
  dueDate?: Date;
  createdAt: Date;
  updatedAt: Date;
}